pyproject.toml
[tool.poetry]
name = "cookie-stand-api"
version = "0.1.0"
description = ""
authors = ["Prabin Singh <pravin544@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.8.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
psycopg2-binary = "^2.9.1"
django-environ = "^0.7.0"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"